feat: adiciona FluentValidation e Result Pattern aos Commands Create e Delete

+++ src/GameHub.Application/Features/Games/Commands/DeleteGame/DeleteGameCommandHandler.cs
@@ -1,22 +1,35 @@
+using FluentValidation;
+using GameHub.Application.Common.Responses;
+using GameHub.Domain.Enums;
 using GameHub.Domain.Repositories;
 using MediatR;
 
 namespace GameHub.Application.Features.Games.Commands.DeleteGame;
 
-public class DeleteGameCommandHandler (IGameRepository gameRepository, IUnitOfWork unitOfWork)
-    : IRequestHandler<DeleteGameCommand, bool>
+public class DeleteGameCommandHandler (IGameRepository gameRepository, IValidator<DeleteGameCommand> validator, IUnitOfWork unitOfWork)
+    : IRequestHandler<DeleteGameCommand, Result>
 {
-    public async Task<bool> Handle(DeleteGameCommand request, CancellationToken cancellationToken)
+    public async Task<Result> Handle(DeleteGameCommand request, CancellationToken cancellationToken)
     {
+        var validationResult = await validator.ValidateAsync(request, cancellationToken);
+        if (!validationResult.IsValid)
+        {
+            var errorMessages = validationResult.Errors
+                .Select(error => error.ErrorMessage)
+                .ToList();
+            
+            return Result.Failure(errorMessages);
+        }
+        
         var game = await gameRepository.GetByIdAsync(request.Id, cancellationToken);
         if (game == null)
         {
-            return false; // Game not found
+            return Result.Failure("Game not found.", ErrorType.NotFound);
         }
         
         gameRepository.Delete(game, cancellationToken);
         await unitOfWork.SaveChangesAsync(cancellationToken);
-        
-        return true; // Deletion successful
+
+        return Result.Success();
     }
 }
